requirements.txt
torch
torchvision
torchaudio
scikit-image
pillow
numpy
matplotlib